Update wp_parse_url() for performance and legacy code removal
With the current supported PHP, it's no longer necessary to use this function's code to mitigate PHP incompatibility. This change aims to remove the legacy fix and improve performance. 

Functions `_get_component_from_parsed_url_array()` and `_wp_translate_php_url_constant_to_key()` are not required for the new version of the function