Tests: Use better domain for HTTP API invalid hostname test.
The previous domain that was used to test for a host whose IPv4 address cannot be resolved, `exampleeeee.com`, got registered and has an A-record now, so it's not invalid anymore.

`.invalid` is intended for use in online construction of domain names that are sure to be invalid and which it is obvious at a glance are invalid.

Reference: [https://datatracker.ietf.org/doc/html/rfc2606#section-2 Reserved Top Level DNS Names: TLDs for Testing, & Documentation Examples].

Follow-up to [52084], [58384], [58388].

Reviewed by peterwilsoncc.
Merges test changes from [59293] to the 6.1 branch.

Props sippis, johnbillion, MattyRob, swissspidy, SergeyBiryukov.
Fixes #62303.




git-svn-id: https://develop.svn.wordpress.org/branches/6.1@59303 602fd350-edb4-49c9-b593-d223f7449a82