Global styles: preload `/global-styles` endpoint responses
This commit fixes a regression whereby requests to global styles endpoints were not being preloaded, resulting in several requests being fired clientside unnecessarily.  

For performance reasons, we  preload the requests so that the data is in the editor store and ready to use straight away.  

The outcome is that the editor loads more quickly.

Follow-up to [62042].

Reviewed by peterwilsoncc.
Merges [59316] to the 6.7 branch.

Props ellatrix, ramonopoly, apermo, peterwilsoncc.
Fixes #62315.



git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59324 602fd350-edb4-49c9-b593-d223f7449a82