Tests/Build tools: Only fail importer tests if plugin is missing.
Reverts an earlier change to the test suite in which the PHPUnit tests could not run if the importer plugin was not available.

This update allows the test suite to run and will fail importer tests if the plugin is not available. 

Follow up to r59085.

Merges [59326] to the 6.7 branch.

Props peterwilsoncc, azaozz.
See #62325.



git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59327 602fd350-edb4-49c9-b593-d223f7449a82