Removed early return
For case -1: was removed return as `-1` of $component is the default value anyway, `null` value is deprecated so it's better to continue the case and return on the end.
I kept typecast `(string) $url` as it's commonly used in WP core code and it's not listed as a bad practice in PHP coding standards. 
Now this should be a performance update, I don't want to declare a new variable like `$string_url=strval($url)` or assign `$url = strval($url);`, from the performance point of view, it's an unnecessary step, this hardly will add to readability.