Feeds: Avoid fatal error with empty `blog_charset` value.
After the SimplePie library was updated to version `1.8.0` in [59141], an edge case has been discovered where a fatal error can encountered if the `blog_charset` option is missing or empty.

In `fetch_feed()`, this option is retrieved using `get_option()` instead of `get_bloginfo( ‘charset’ )`. The latter will detect this scenario and apply a default value of `UTF-8` and is already used interchangeably throughout Core. This switches to `get_bloginfo( ‘charset’ )` instead to prevent this edge case.

Reviewed by davidbaumwald.
Merges [59382] to the 6.7 branch.

Props david.binda, davidbaumwald, SergeyBiryukov, sabernhardt, azaozz, peterwilsoncc.
Fixes #62354.

git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59383 602fd350-edb4-49c9-b593-d223f7449a82