Only prime salt options early once.
This adds a check to see if `wp_salt()` related options have already been primed before attempting to prime them again. This is necessary because the `wp_salt()` function can be called multiple times during the bootstrap process, and we need to ensure that the options are only primed once.

This also avoids an error that can occur if the salt related constants are defined after the first time `wp_salt()` is called, in which case the salt will not be present in the `$duplicated_keys` array.