Media: Avoid running expensive logic twice using GD.
Support for uploading AVIF was added in [57524]. A new block of conditional logic was added determine which function should be used to create the new image file that resulted in these expensive functions being run twice.

This combines the two conditional logic to ensure the appropriate function is only run once regardless of format.

Reviewed by adamsilverstein.
Merges [59413] to the 6.7 branch.

Props adamsilverstein, glynnquelch.
Fixes #62331.

git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59423 602fd350-edb4-49c9-b593-d223f7449a82