Bundled Theme: Pin a theme.json schema version to Twenty Twenty-Five.
Each theme’s theme.json schema version should be pinned to the version that was valid at the time it was released. This means all JSON files in the theme should be pinned to the appropriate schema version, not just theme.json.

Reviewed by jorbin.
Merges [59448] and [59451] to the 6.7 branch.

Props im3dabasia1, poena, mukesh27, desrosj.
Fixes #62455.



git-svn-id: https://develop.svn.wordpress.org/branches/6.7@59546 602fd350-edb4-49c9-b593-d223f7449a82